+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.0] - 2017-07-31
+### Added
+- Method Injector::merge()
+- Method Injector::getLoggerHelper()
+
+### Changed
+- LoggerHelper implement PSR LoggerAwareInterface
+
 ## [1.2.0] - 2017-06-11 
 ### Changed
 - Allow use value without key with interface name: for instances will be used class, for registered types -- self class name.